rename default_qubit file to default_qubit_legacy
