+++ pennylane/devices/default_qubit_legacy.py
