+++ pennylane/devices/default_qutrit.py
@@ -24,7 +24,7 @@ import numpy as np
 import pennylane as qml  # pylint: disable=unused-import
 from pennylane import QutritDevice, QutritBasisState, DeviceError
 from pennylane.wires import WireError
-from pennylane.devices.default_qubit import _get_slice
+from pennylane.devices.default_qubit_legacy import _get_slice
 from .._version import __version__
 
 # tolerance for numerical errors
